# docs/swagger-json.json
{"openapi":"3.0.0","info":{"title":"Shippings Swagger Docs","version":"1.0.0","description":"Swagger docs for shippings project"},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"LoginRequest":{"type":"object","required":["usernameOrEmail","password"],"properties":{"usernameOrEmail":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","required":["token","user"],"properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}}},"LoginResponseTokenOnly":{"type":"object","required":["token"],"properties":{"token":{"type":"string"}}},"CreateOrder":{"type":"object","required":["senderId","receiverId","productCategory","weightGrams","targetAddress","productDescription","dimensionX","dimensionY","dimensionZ","quantity","sourceCityId","targetCityId"],"properties":{"senderId":{"type":"integer","description":"The ID of the sender."},"receiverId":{"type":"integer","description":"The ID of the receiver."},"productCategory":{"type":"string","description":"The product category."},"weightGrams":{"type":"string","description":"The weight of the product in grams."},"targetAddress":{"type":"string","description":"The target address for the delivery."},"productDescription":{"type":"string","description":"A brief description of the product."},"dimensionX":{"type":"number","format":"float","description":"The dimension of the product along the X-axis."},"dimensionY":{"type":"number","format":"float","description":"The dimension of the product along the Y-axis."},"dimensionZ":{"type":"number","format":"float","description":"The dimension of the product along the Z-axis."},"quantity":{"type":"integer","description":"The quantity of the product."},"sourceCityId":{"type":"integer","description":"The ID of the source city."},"targetCityId":{"type":"integer","description":"The ID of the target city."}},"description":"DTO for creating an order without optional or ID fields."},"CreateOrderResponse":{"type":"object","required":["order"],"properties":{"order":{"$ref":"#/components/schemas/Order"}},"description":"DTO for the response of creating an order. Contains the created order."},"AssignOrder":{"type":"object","required":["orderShortId","transporterId","routeId","expectedReachData","vehicleId"],"properties":{"orderShortId":{"type":"string","description":"A short identifier for the order."},"transporterId":{"type":"integer","format":"int64","description":"ID of the transporter assigned to the order."},"routeId":{"type":"integer","format":"int64","description":"ID of the route the order will follow."},"expectedReachData":{"type":"string","format":"date-time","description":"Expected date and time the order should reach its destination."},"vehicleId":{"type":"integer","format":"int64","description":"ID of the vehicle assigned for the order."}},"description":"DTO for updating an existing order with transport, route, and vehicle info."},"AssignOrderResponse":{"type":"object","required":["order"],"properties":{"order":{"$ref":"#/components/schemas/Order"}},"description":"DTO for the response of assigning an order to a transporter. Contains the created order."},"QueryOrderResponse":{"type":"object","required":["order"],"properties":{"order":{"$ref":"#/components/schemas/Order"}},"description":"DTO for the response of querying an order by short ID."},"StatisticsOrder":{"type":"object","properties":{"meanTime":{"type":"number"},"minTime":{"type":"number"},"maxTime":{"type":"number"},"count":{"type":"number"}}},"StatisticsOrderByTransporter":{"type":"object","properties":{"transporter":{"$ref":"#/components/schemas/Transporter"},"statistics":{"$ref":"#/components/schemas/StatisticsOrder"}}},"StatisticsOrderByCity":{"type":"object","properties":{"city":{"$ref":"#/components/schemas/City"},"statistics":{"$ref":"#/components/schemas/StatisticsOrder"}}},"CreateUserRequest":{"type":"object","required":["username","document","documentType","password","email","status","firstName","firstSurname"],"properties":{"username":{"type":"string"},"document":{"type":"string"},"documentType":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"status":{"type":"string"},"firstName":{"type":"string"},"secondName":{"type":"string","nullable":true},"firstSurname":{"type":"string"},"secondSurname":{"type":"string","nullable":true},"roles":{"type":"array","items":{"type":"string"}}},"description":"The data transfer object used to create a new user"},"CreateUserResponse":{"type":"object","required":["user","auth"],"properties":{"user":{"$ref":"#/components/schemas/User"},"auth":{"$ref":"#/components/schemas/LoginResponseTokenOnly"}},"description":"Response of user creation process"},"City":{"type":"object","required":["id","name"],"properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}},"description":"Represents a city with an ID and name."},"Order":{"type":"object","required":["id","senderId","receiverId","productCategory","weightGrams","targetAddress","productDescription","dimensionX","dimensionY","dimensionZ","status","sourceCityId","targetCityId","latitude","longitude","solvedAddress","solvedCity"],"properties":{"id":{"type":"integer"},"senderId":{"type":"integer"},"receiverId":{"type":"integer"},"productCategory":{"type":"string"},"weightGrams":{"type":"number"},"targetAddress":{"type":"string"},"productDescription":{"type":"string"},"dimensionX":{"type":"number"},"dimensionY":{"type":"number"},"dimensionZ":{"type":"number"},"quantity":{"type":"number"},"status":{"type":"string"},"transporterId":{"type":"integer","nullable":true},"routeId":{"type":"integer","nullable":true},"sourceCityId":{"type":"integer"},"targetCityId":{"type":"integer"},"latitude":{"type":"number"},"longitude":{"type":"number"},"solvedAddress":{"type":"string"},"solvedCity":{"type":"string"},"assignedDate":{"type":"string","format":"date-time"},"expectedReachDate":{"type":"string","format":"date-time"},"actualReachDate":{"type":"string","format":"date-time"},"shortId":{"type":"string"}},"description":"An order model that includes logistics, location, and transportation data"},"Privilege":{"type":"object","required":["id","userId","name"],"properties":{"id":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"name":{"type":"string"}},"description":"Represents a privilege assigned to a user."},"Route":{"type":"object","required":["id","name"],"properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}},"description":"Represents a route through which deliveries or transportation is made."},"Transporter":{"type":"object","required":["id","userId","status"],"properties":{"id":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"status":{"type":"string"}},"description":"Represents a transporter, which is associated with a user and has a status."},"User":{"type":"object","required":["username","document","documentType","password","email","status","firstName","firstSurname"],"properties":{"username":{"type":"string"},"document":{"type":"string"},"documentType":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"status":{"type":"string"},"firstName":{"type":"string"},"secondName":{"type":"string","nullable":true},"firstSurname":{"type":"string"},"secondSurname":{"type":"string","nullable":true},"roles":{"type":"array","items":{"type":"string"}}},"description":"A user model containing basic information"},"Vehicle":{"type":"object","required":["id","name","capacity"],"properties":{"id":{"type":"integer","description":"The unique identifier for the vehicle"},"name":{"type":"string","description":"The name of the vehicle"},"capacity":{"type":"number","format":"float","description":"The capacity of the vehicle"}},"description":"A model representing a vehicle with an ID, name, and capacity."}}},"security":[{"bearerAuth":[]}],"paths":{"/auth":{"post":{"summary":"Login to obtain an access token","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"Logged in successfully"},"400":{"description":"Invalid data passed"},"401":{"description":"Invalid Credentials"}}}},"/docs":{"get":{"summary":"Watch swagger API docs","tags":["Swagger"],"responses":{"200":{"description":"Retrieved swagger logs successfully"}}}},"/order":{"post":{"summary":"Create a new shipment order","tags":["Order"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrder"}}}},"responses":{"201":{"description":"Order created successfully"},"400":{"description":"Invalid data passed"}}}},"/order/{shortId}/assign":{"post":{"summary":"Assign a new shipment order","tags":["Order"],"parameters":[{"in":"path","name":"shortId","required":true,"schema":{"type":"string"},"description":"Short ID of order to assign"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignOrder"}}}},"responses":{"200":{"description":"Order assigned successfully"},"400":{"description":"Invalid data passed"}}}},"/order/{shortId}":{"get":{"summary":"Get shipment order state","tags":["Order"],"parameters":[{"in":"path","name":"shortId","required":true,"schema":{"type":"string"},"description":"Short ID of order to assign"}],"responses":{"200":{"description":"Order assigned successfully"},"400":{"description":"Invalid data passed"}}}},"/order/statistics":{"get":{"summary":"Get shipment order statistics","tags":["Order"],"parameters":[{"in":"query","name":"minOrderAssignedDate","required":true,"schema":{"type":"string","format":"date"},"description":"Min Order Assigned Date"},{"in":"query","name":"maxOrderAssignedDate","required":true,"schema":{"type":"string","format":"date"},"description":"Max Order Assigned Date"},{"in":"query","name":"status","required":true,"schema":{"type":"string"},"description":"Status to Filter"},{"in":"query","name":"groupBy","required":true,"schema":{"type":"string"},"description":"Field to group by"},{"in":"query","name":"limit","required":true,"schema":{"type":"integer"},"description":"Pagination max size"},{"in":"query","name":"page","required":true,"schema":{"type":"integer"},"description":"Pagination current page"},{"in":"query","name":"transporterId","required":true,"schema":{"type":"integer"},"description":"Transporter ID to filter"}],"responses":{"200":{"description":"Statistics generated successfully","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatisticsOrder"},{"$ref":"#/components/schemas/StatisticsOrderByTransporter"},{"$ref":"#/components/schemas/StatisticsOrderByCity"}]}}}},"400":{"description":"Invalid data passed"}}}},"/user/register":{"post":{"summary":"Register to create a new user","tags":["User"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}}},"responses":{"201":{"description":"User created successfully"},"400":{"description":"Invalid data passed"},"409":{"description":"User already exists"}}}}},"tags":[]}